refactor player controls into new component
